pass cspxX

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: honguyen <honguyen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/27 15:59:02 by honguyen          #+#    #+#             */
-/*   Updated: 2023/12/01 19:56:53 by honguyen         ###   ########.fr       */
+/*   Updated: 2023/12/01 20:00:32 by honguyen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ int		len_s_base(long long n, int base);
 int		len_x(unsigned int n, t_formats formats);
 int		totalize_len(int n, int no_digit, t_formats *p);
 int		print_sign(t_formats formats, int n);
-void	ft_putnbr(int n, int *p_np);
+void	ft_putnbr(long n, int *p_np);
 void	print_hex(unsigned long n, char x, t_formats formats, int *np);
 int		print_0x(t_formats formats, char c, unsigned int x);
 #endif
